pyproject.toml
[project]
name = "fastapi-learn"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13.2"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "fastapi[all]>=0.116.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
